requirements.txt
flet>=0.21.0
ollama>=0.1.6
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
